Thank You Page Added
This is the page where it brings you after you've donated x amount of money. There is a random number generator that displays the total amount of money from everyone that has donated. There is also a button at the bottom that will bring you back to the main page.