Added Payment Amount Activity & Confirmation Fragment
It brings you to a new activity after you fill out all of the card information where you put in the amount of money you would like to pay. When you click the PAY button, a fragment will pop up asking the user to confirm the payment. I also made the app a lot cleaner, commented all the code, and readjusted all the widgets that needed some work.